+++ CHANGELOG.md
@@ -290,7 +290,7 @@
 
 ## 1.66.1 - 2023-06-07
 
-- Update utils.ts (#686)
+- Update index.ts (#686)
 
 ## 1.66.0 - 2023-06-06
