+++ src/__tests__/send-request.ts
@@ -4,7 +4,7 @@ import { addParamsToURL, encodePostData, xhr } from '../send-request'
 import { assert, boolean, property, uint8Array, VerbosityLevel } from 'fast-check'
 import { Compression, PostData, XHROptions, XHRParams } from '../types'
 
-import { _isUndefined } from '../type-utils'
+import { _isUndefined } from '../utils/type-utils'
 
 jest.mock('../config', () => ({ DEBUG: false, LIB_VERSION: '1.23.45' }))
