+++ src/autocapture-utils.ts
@@ -6,7 +6,7 @@
 import { AutocaptureConfig } from 'types'
 import { _each, _includes, _trim, logger } from './utils'
 
-import { _isNull, _isString, _isUndefined } from './type-utils'
+import { _isNull, _isString, _isUndefined } from './utils/type-utils'
 
 export function getClassName(el: Element): string {
     switch (typeof el.className) {
