+++ src/extensions/exception-autocapture/index.ts
@@ -4,7 +4,7 @@ import { DecideResponse, Properties } from '../../types'
 import { ErrorEventArgs, ErrorProperties, errorToProperties, unhandledRejectionToProperties } from './error-conversion'
 import { isPrimitive } from './type-checking'
 
-import { _isArray, _isObject, _isUndefined } from '../../type-utils'
+import { _isArray, _isObject, _isUndefined } from '../../utils/type-utils'
 
 const EXCEPTION_INGESTION_ENDPOINT = '/e/'
