+++ src/request-queue.ts
@@ -2,7 +2,7 @@ import { RequestQueueScaffold } from './base-request-queue'
 import { _each } from './utils'
 import { Properties, QueuedRequestData, XHROptions } from './types'
 
-import { _isUndefined } from './type-utils'
+import { _isUndefined } from './utils/type-utils'
 
 export class RequestQueue extends RequestQueueScaffold {
     handlePollRequest: (url: string, data: Properties, options?: XHROptions) => void
